# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Controller",
        "description": "",
        "version": "1.0.0"
    },
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "get list product",
                "operationId": "App\\Http\\Controllers\\Api\\ProductController::getAllProduct",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "price_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "access denied"
                    },
                    "409": {
                        "description": "exists"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "get product detail",
                "operationId": "App\\Http\\Controllers\\Api\\ProductController::getProductDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "access denied"
                    },
                    "409": {
                        "description": "exists"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/slider": {
            "get": {
                "tags": [
                    "slider"
                ],
                "summary": "get list slider",
                "operationId": "App\\Http\\Controllers\\Api\\SliderController::getSliderbyPosition",
                "parameters": [
                    {
                        "name": "position",
                        "in": "query",
                        "description": "1: Banner Large , 2: Banner Small",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "access denied"
                    },
                    "409": {
                        "description": "exists"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}